replaced uint8_t with uint16_t

+++ src/config.h
@@ -9,10 +9,10 @@ constexpr bool SIMULATE_HARDWARE = false;
 constexpr bool SIMULATE_HARDWARE = true;
 #endif
 
-constexpr uint8_t SCREEN_HEIGHT = 480;
-constexpr uint8_t SCREEN_WIDTH = 800;
+constexpr uint16_t SCREEN_HEIGHT = 480;
+constexpr uint16_t SCREEN_WIDTH = 800;
 
-constexpr uint8_t SEGMENT_HEIGHT = 20;
-constexpr uint8_t SEGMENT_WIDTH = 20;
+constexpr uint16_t SEGMENT_HEIGHT = 20;
+constexpr uint16_t SEGMENT_WIDTH = 20;
 
 #endif // CONFIG_H_
