+++ src/drivers/tty_display.h
@@ -9,8 +9,8 @@ class TtyDisplay : public Display
 public:
     TtyDisplay(bool verticalOrientation);
     int clear() override;
-    virtual void drawText(const std::string& text, const uint8_t& x, const uint8_t& y, bool isHighlighted) override;
-    virtual void drawBox(const uint8_t& width, const uint8_t& height, const bool hasBorder, const bool isFilled, const uint8_t& x, const uint8_t& y) override; 
+    virtual void drawText(const std::string& text, const uint16_t& x, const uint16_t& y, bool isHighlighted) override;
+    virtual void drawBox(const uint16_t& width, const uint16_t& height, const bool hasBorder, const bool isFilled, const uint16_t& x, const uint16_t& y) override; 
     virtual void refresh() override {};
 };
