+++ src/ui/screen.h
@@ -10,12 +10,12 @@ class Display;
 class Screen
 {
 private:
-    uint8_t m_width;
-    uint8_t m_height;
+    uint16_t m_width;
+    uint16_t m_height;
     std::unique_ptr<Widget> m_rootWidget; //Make a subclass called RootWidget
 public:
     Screen() = default;
-    Screen(uint8_t width, uint8_t height);
+    Screen(uint16_t width, uint16_t height);
     Screen(const Screen& other);
     Screen& operator=(const Screen& other);
     void setRootWidget(std::unique_ptr<Widget> widget) {m_rootWidget = std::move(widget);}
