+++ src/ui/vertical_widget.h
@@ -10,7 +10,7 @@ class VerticalWidget : public Widget
 private:
 	std::vector<std::unique_ptr<Widget>> m_widgets;
 public:
-	virtual void render(Display* display, const uint8_t& x, const uint8_t& y) const override;
+	virtual void render(Display* display, const uint16_t& x, const uint16_t& y) const override;
 	virtual std::unique_ptr<Widget> clone() override;
 	void add(std::unique_ptr<Widget>& widget);
 };
